Separate static-Ranger games into ranger_games/
The game_engine root and scripting/ folder mixed three different concepts:
loadable Wasm/TSX games (games/), the engine core runtime (scripting/), and
games whose logic is static Ranger source compiled to a native binary. The
last group was scattered — hand-written Pong at the root, AOT native-compiled
games (pong/pacman/invaders, emitted from TSX) buried in scripting/ next to the
runtime.

Collect all static-Ranger games into a clearly named gallery/game_engine/
ranger_games/ folder ("Ranger as a platform"), leaving scripting/ as engine
core only and games/ as the loadable-game catalog.

- Move both families into ranger_games/: hand-written pong{,_core,_render,_sdl}
  + sprite_char_sdl + streaming_world_sdl, and the AOT native games/runners
  (pong/pacman/invaders/spawner/counter). The committed ES6 reference build
  (pong.js) moves with its source.
- The reusable native host (native_game_bridge, game_sdl_native_host,
  game_native_runtime) and the render shims (framebuffer, gfx_sdl, wasm_runtime)
  are engine core and stay put; moved files now import them via ../scripting/
  and ../.
- Update every reference in lockstep: package.json (engine:*), the SDL/native/
  Pi build scripts, tests (ts-to-ranger-*, game-engine-render, pong_render_probe
  fixture), and .gitignore.
- Docs/skills updated to the three-way split: new ranger_games/README.md, plus
  README.md, AGENTS.md and scripting/AGENTS.md.

Verified by compiling every moved entrypoint to ES6 and running the native Pong
runner headless (180 frames) through the updated paths.
