+++ gallery/game_engine/ranger_games/invaders_native_sdl_runner.rgr
@@ -9,7 +9,7 @@
 ;   npm run engine:game-sdl-native:run:invaders
 ; ==============================================================================
 
-Import "./game_sdl_native_host.rgr"
+Import "../scripting/game_sdl_native_host.rgr"
 Import "./invaders_native_game.rgr"
 
 class InvadersNativeSdlRunner {
