+++ gallery/game_engine/ranger_games/pong_core.rgr
